On remplace par du double, en fait. (précision)

+++ ag41_project_sample1/Client.h
@@ -14,26 +14,24 @@
 
 using namespace std;
 
-namespace Donnees
-{
-	class Client
-	{
-	private:
-		vector<Commande*> *listeCommande;
-		string nom;
-		float kStockage;
-	public:
-		Commande* derniereCommande() const;
-		Commande* premiereCommande() const;
-		Client(string s);
-		vector<Commande*>* getCommande();
-		void addCommande(Commande* c);
-		virtual ~Client();
-		string getNom() const;
-		void setNom(string s);
-		friend ostream& operator<< (ostream& flux, Client& c);
-		float getKStockage();
-		void setKStockage(float k);
-	};
+namespace Donnees {
+class Client {
+private:
+	vector<Commande*> *listeCommande;
+	string nom;
+	double kStockage;
+public:
+	Commande* derniereCommande() const;
+	Commande* premiereCommande() const;
+	Client(string s);
+	vector<Commande*>* getCommande();
+	void addCommande(Commande* c);
+	virtual ~Client();
+	string getNom() const;
+	void setNom(string s);
+	friend ostream& operator<<(ostream& flux, Client& c);
+	double getKStockage();
+	void setKStockage(double k);
+};
 }
 #endif /* CLIENT_H_ */
