+++ ag41_project_sample1/Modification.cpp
@@ -9,14 +9,14 @@
 
 namespace Calcul {
 
-Modification::Modification(Action* ac, int tDep, int tArr): act(ac), tDepart(tDep), tArrive(tArr) {
+Modification::Modification(Action* ac, double tDep, double tArr) :
+		act(ac), tDepart(tDep), tArrive(tArr) {
 }
 
 Modification::~Modification() {
 }
 
-bool Modification::operator==(Modification &m)
-{
+bool Modification::operator==(Modification &m) {
 	bool b = false;
 	if (tDepart == m.tDepart && tArrive == m.tArrive)
 		if (act == m.act)
